Fix constant redefinition warning
In e18fe2a the Jbuilder::VERSION constant was introduced but in 9aa3dd9
it was used in the gemspec which changed the loading order so that the
version constant was loaded first. This defined Jbuilder as an Object
subclass rather than the intended BasicObject and when jbuilder/jbuilder
was required it redefined the Jbuilder constant and obliterates the
VERSION constant. This commit ensures that the version constant exists
and the Jbuilder parent class is BasicObject.